Boolean started, root URL response, URL entries, 404s
The boolean search has been written and appears to work, but tests need 
to be written and it needs to be documented. Note that '|', '&' and ' '
need to be url encoded.

A root URL response has been created, with links to start exploring. Tests
have been written for that, but it still needs documenting.

A URL entry has been added to all entries in the response. This needs to
be documented, and maybe extended to always work.

Exceptions are now slightly better written.

A 404 will be returned if an asset is requested that does not exist.


git-svn-id: https://svn.aptivate.org/svn/idsapi/dev@5191 679a26be-e637-0410-a801-b7d8065a4221