Started registration and refactored assets/categories stuff
This commit is primarily for the assets/categories split. This includes
moving to object_id as the key id field, and having an initial letter on
it eg 1234 becomes A1234 (or C1234 - they both exist). You can now search
across all assets, but not all objects.

There is also some registration work in here, but that is not finished yet.


git-svn-id: https://svn.aptivate.org/svn/idsapi/dev@5537 679a26be-e637-0410-a801-b7d8065a4221