Don't escape whitespace when building SOLR query
It doesn't seem to be needed so I've overridden the list of special chars given
in the sunburnt library - the SOLR docs also suggest otherwise, so either
sunburnt is wrong or I don't understand the reason why the library developer
chose to escape whitespace.

This fixes the issue of '\' being inserted in spaces in SOLR queries.



git-svn-id: https://svn.aptivate.org/svn/idsapi/dev@10151 679a26be-e637-0410-a801-b7d8065a4221