Move field list handling to DataMunger
Due to all the wildcard fields, we can't ask Solr to only return a
subset of fields, so we have to get all the fields back, and then
exclude them when we assemble the return data.

However we still check the extra_fields parameter for illegal fields
before we send the query to Solr.

Also did some view refactoring.

git-svn-id: https://svn.aptivate.org/svn/idsapi/dev@11164 679a26be-e637-0410-a801-b7d8065a4221