Ignore more fields in DataMunger
fields called XX_search_api_* are byproducts of the drupal importer, we
can ignore them.

And fields that contain _search_hub_xx or _sort_hub_xx are used for
sorting/searching, but are copies of other fields.  So we don't show
them to the end user.

Also, fields that start with "hub_" shouldn't be split up by rules.  For
now we show them with the "hub_" prefix, but that might change.

git-svn-id: https://svn.aptivate.org/svn/idsapi/dev@11202 679a26be-e637-0410-a801-b7d8065a4221