reset item_id to object_id (and _type), lang_pref, source_pref
for object_id, we're using object_id_hub_zz from the Solr index (which
is a copy of item_id)

Also add the lang_pref and source_pref search parameters.

And a few other changes.  Not confident it's all working yet, but need
to put the fixes I have up on the server