Use site packages for mysqlclient 1.3.10 (dye)
This not really an ideal situation, we can't install the MySQL C client
libraries on the production host, so can't install the latest version of
mysqlclient (which seems to be needed for Django >1.11 support, and to
fix a load of bugs).

So for the Django app I think we're going to have to use
mysql-connector-python instead.

However DYE is still tied to a version of mysqlclient (less critical),
so will use out dated version that comes with the platform in the site
packages (/chroot/python36/lib/python3.6/site-packages/).

If this doesn't work out we can remove this dependency from DYE / make
it optional.