Apache config for both SSL and non-SSL sites
We need to keep non-SSL going for legacy API clients (traffic isn't
considered a risk to have in plain text).

Though allowed in the config spec it seems a VirtualHost def with both ports no
longer works as expected (will try to serve SSL on both), so here we break the
config into two VHOSTS sharing an include.

Using a common shared WSGIDaemonProcess so resources are not wasted.